pyproject.toml
[tool.poetry]
name = "nolabsserver"
version = "0.1.0"
description = ""
authors = ["JaktensTid <jaktenstid1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.25.0"
fastapi = "^0.109.0"
numpy = "^1.26.3"
requests = "^2.31.0"
localisation-microservice = {path = "microservices/localisation/client"}
obonet = "^1.0.0"
setuptools = "^69.0.3"
gene-ontology-microservice = {path = "microservices/gene_ontology/client"}
rdkit = "^2023.9.4"
protein-design-microservice = {path = "microservices/protein_design/client"}
conformations-microservice = {path = "microservices/conformations/client"}
python-json-logger = "^2.0.7"
python-dateutil = "^2.8.2"
python-slugify = "^8.0.2"
python-multipart = "^0.0.6"
msa-light-microservice = {path = "microservices/msa_light/client"}
p2rank-microservice = {path = "microservices/p2rank/client"}
solubility-microservice = {path = "microservices/solubility/client"}
jupyter = "^1.0.0"
umol-microservice = {path = "microservices/umol/client"}
esmfold-light-microservice = {path = "microservices/esmfold_light/client"}
esmfold-microservice = {path = "microservices/esmfold/client"}
diffdock-microservice = {path = "microservices/diffdock/client"}
biobuddy-microservice = {path = "microservices/biobuddy/client"}
pubmed-query-microservice = {path = "microservices/pubmed_query/client"}
rcsb-pdb-query-microservice = {path = "microservices/rcsb_pdb_query/client"}
chembl-query-microservice = {path = "microservices/chembl_query/client"}
aiohttp = "^3.9.3"
rosettafold-microservice = {path = "microservices/rosettafold/client"}
leaf-basedlabs = "^0.0.7"
reinvent-microservice = {path = "microservices/reinvent/client"}

[project]
dependencies = [
    "fastapi",
    "mypy",
    "pydantic",
    "obonet",
    "jsonpickle",
    "python-multipart",
    "rdkit"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
